[refactoring] Move most protobuf enums into messages
Having enums on the root context has some undesired behavior in JavaScript.
It is also nicer in Go, when wrapped in a message.
The Team enum did not fit in any message well, though...

+++ internal/app/statemachine/change_config.go
@@ -20,7 +20,7 @@ func (s *StateMachine) UpdateConfig(newState *state.State, change *UpdateConfig)
 	}
 	for eventType, behavior := range change.GameEventBehavior {
 		log.Printf("Change game event behavior for %v to %v", eventType, behavior)
-		newState.GameEventBehavior[int32(eventType)] = behavior
+		newState.GameEventBehavior[eventType] = behavior
 	}
 	return
 }
